+++ packages/core/src/service/Config.ts
@@ -2,7 +2,7 @@ import rfdc from 'rfdc'
 import type { ExternalEventEmitter } from '../common/index.js'
 import { Arrayable, bufferToString, merge, TypePredicates } from '../common/index.js'
 import { ErrorSeverity } from '../source/index.js'
-import { FileCategories, RegistryCategories } from '../symbol/index.js'
+import { DataFileCategories, FileCategories, RegistryCategories } from '../symbol/index.js'
 import type { Project } from './Project.js'
 /* eslint-disable no-restricted-syntax */
 
@@ -115,6 +115,12 @@ export interface EnvConfig {
 	>
 	permissionLevel: 1 | 2 | 3 | 4
 	plugins: string[]
+	/**
+	 * Whether to enable caching of mcdoc simplified types.
+	 *
+	 * May become corrupt after changing game versions, so this is currently disabled by default.
+	 */
+	enableMcdocCaching: boolean
 }
 
 export type LinterSeverity = 'hint' | 'information' | 'warning' | 'error'
@@ -319,7 +325,7 @@ export namespace SymbolLinterConfig {
 export const VanillaConfig: Config = {
 	env: {
 		dataSource: 'GitHub',
-		dependencies: ['@vanilla-datapack', '@vanilla-mcdoc'],
+		dependencies: ['@vanilla-datapack', '@vanilla-resourcepack', '@vanilla-mcdoc'],
 		exclude: ['@gitignore', '.vscode/', '.github/'],
 		customResources: {},
 		feature: {
@@ -352,6 +358,7 @@ export const VanillaConfig: Config = {
 		permissionLevel: 2,
 		plugins: [],
 		mcmetaSummaryOverrides: {},
+		enableMcdocCaching: false,
 	},
 	format: {
 		blockStateBracketSpacing: { inside: 0 },
@@ -410,9 +417,13 @@ export const VanillaConfig: Config = {
 			{
 				if: [
 					{ category: RegistryCategories, namespace: 'minecraft' },
-					{ category: [...FileCategories, 'bossbar', 'objective', 'team'] },
+					{ category: [...DataFileCategories, 'bossbar', 'objective', 'team'] },
 				],
 				then: { report: 'warning' },
+			},
+			{
+				if: { category: ['attribute_modifier', 'attribute_modifier_uuid', 'tag'] },
+				then: { declare: 'public' },
 			},
 			{
 				then: { declare: 'block' },