+++ packages/java-edition/src/binder/index.ts
@@ -20,21 +20,28 @@ interface Resource {
 	path: string
 	category: FileCategory
 	ext: `.${string}`
+	pack: 'data' | 'assets'
+	identifier?: string
 	since?: ReleaseVersion
 	until?: ReleaseVersion
 }
 
-const Resources = new Map<string, Resource>()
+const Resources = new Map<string, Resource[]>()
 
 function resource(path: string, resource: Partial<Resource> & { category: FileCategory }): void
 function resource(path: FileCategory, resource: Partial<Resource>): void
 function resource(path: string, resource: Partial<Resource> = {}) {
-	Resources.set(path, {
-		path,
-		category: resource.category ?? path as FileCategory,
-		ext: resource.ext ?? '.json',
-		...resource,
-	})
+	const previous = Resources.get(path) ?? []
+	Resources.set(path, [
+		...previous,
+		{
+			path,
+			category: resource.category ?? path as FileCategory,
+			ext: resource.ext ?? '.json',
+			pack: resource.pack ?? 'data',
+			...resource,
+		},
+	])
 }
 
 // Pre-1.21 data pack plurals
@@ -120,6 +127,42 @@ for (const registry of TaggableResourceLocationCategories) {
 	resource(`tags/${registry}`, { category: `tag/${registry}`, since: '1.18' })
 }
 
+// Resource pack
+resource('atlases', { pack: 'assets', category: 'atlas', since: '1.19.3' })
+resource('blockstates', { pack: 'assets', category: 'block_definition' })
+resource('equipment', { pack: 'assets', since: '1.21.4' })
+resource('font', { pack: 'assets', since: '1.16' })
+resource('font', { pack: 'assets', category: 'font/ttf', since: '1.16', ext: '.ttf' })
+resource('font', { pack: 'assets', category: 'font/otf', since: '1.16', ext: '.otf' })
+resource('font', { pack: 'assets', category: 'font/unihex', since: '1.20', ext: '.zip' })
+resource('items', { pack: 'assets', category: 'item_definition', since: '1.21.4' })
+resource('lang', { pack: 'assets' })
+resource('models', { pack: 'assets', category: 'model' })
+resource('models/equipment', {
+	pack: 'assets',
+	category: 'equipment',
+	since: '1.21.2',
+	until: '1.21.4',
+})
+resource('particles', { pack: 'assets', category: 'particle' })
+resource('post_effect', { pack: 'assets', since: '1.21.2' })
+resource('shaders/post', { pack: 'assets', category: 'post_effect', until: '1.21.2' })
+resource('shaders', { pack: 'assets', category: 'shader' })
+resource('shaders', { pack: 'assets', category: 'shader/fragment', ext: '.fsh' })
+resource('shaders', { pack: 'assets', category: 'shader/vertex', ext: '.vsh' })
+resource('sounds', { pack: 'assets', category: 'sound', ext: '.ogg' })
+resource('textures', { pack: 'assets', category: 'texture', ext: '.png' })
+resource('textures', { pack: 'assets', category: 'texture_meta', ext: '.png.mcmeta' })
+
+resource('lang', { pack: 'assets', category: 'lang/deprecated', identifier: 'deprecated' })
+resource('', { pack: 'assets', category: 'sounds', identifier: 'sounds' })
+resource('', {
+	pack: 'assets',
+	category: 'regional_compliancies',
+	identifier: 'regional_compliancies',
+})
+resource('', { pack: 'assets', category: 'gpu_warnlist', identifier: 'gpu_warnlist' })
+
 export function* getRels(
 	uri: string,
 	rootUris: readonly RootUriString[],
@@ -128,7 +171,7 @@ export function* getRels(
 
 	const parts = uri.split('/')
 	for (let i = parts.length - 2; i >= 0; i--) {
-		if (parts[i] === 'data') { // TODO: support assets
+		if (parts[i] === 'data' || parts[i] === 'assets') {
 			yield parts.slice(i).join('/')
 		}
 	}
@@ -145,34 +188,62 @@ export function dissectUri(uri: string, ctx: UriBinderContext) {
 
 	for (const rel of rels) {
 		const parts = rel.split('/')
-		if (parts.length < 4) {
+		if (parts.length < 3) {
 			continue
 		}
 		const [pack, namespace, ...rest] = parts
-		if (pack !== 'data') {
-			continue // TODO: support assets
+		if (pack !== 'data' && pack !== 'assets') {
+			continue
 		}
-		let resource: Resource | undefined = undefined
-		let matchIndex = 0
-		for (let i = 1; i < rest.length; i += 1) {
-			const res = Resources.get(rest.slice(0, i).join('/'))
-			if (res) {
-				resource = res
-				matchIndex = i
+		const candidateResources: [Resource, string][] = []
+		if (rest.length === 1) {
+			const resources = Resources.get('')
+			for (const res of resources ?? []) {
+				if (res.pack !== pack) {
+					continue
+				}
+				let identifier = rest[0]
+				if (!identifier.endsWith(res.ext)) {
+					continue
+				}
+				identifier = identifier.slice(0, -res.ext.length)
+				if (res.identifier && identifier !== res.identifier) {
+					continue
+				}
+				candidateResources.push([res, identifier])
 			}
 		}
-		if (!resource) {
+		for (let i = 1; i < rest.length; i += 1) {
+			const resources = Resources.get(rest.slice(0, i).join('/'))
+			for (const res of resources ?? []) {
+				if (res.pack !== pack) {
+					continue
+				}
+				let identifier = rest.slice(i).join('/')
+				if (!identifier.endsWith(res.ext)) {
+					continue
+				}
+				identifier = identifier.slice(0, -res.ext.length)
+				if (res.identifier && identifier !== res.identifier) {
+					continue
+				}
+				candidateResources.push([res, identifier])
+			}
+		}
+		if (candidateResources.length === 0) {
 			continue
 		}
-		let identifier = rest.slice(matchIndex).join('/')
-		if (!identifier.endsWith(resource.ext)) {
-			continue
+		// Finding the last, because that will be the deepest match
+		let res = candidateResources.findLast(([res]) => matchVersion(release, res.since, res.until))
+		if (res !== undefined) {
+			return { ok: true, ...res[0], namespace, identifier: res[1], expected: undefined }
 		}
-		identifier = identifier.slice(0, -resource.ext.length)
-		if (!matchVersion(release, resource.since, resource.until)) {
-			let expected: string | undefined = undefined
-			for (const [path, other] of Resources) {
-				if (other.category !== resource.category) {
+		// Try to find the expected path that matches the current version
+		res = candidateResources[candidateResources.length - 1]
+		let expected: string | undefined = undefined
+		for (const [path, others] of Resources) {
+			for (const other of others) {
+				if (other.category !== res[0].category) {
 					continue
 				}
 				if (matchVersion(release, other.since, other.until)) {
@@ -180,9 +251,8 @@ export function dissectUri(uri: string, ctx: UriBinderContext) {
 					break
 				}
 			}
-			return { ok: false, ...resource, namespace, identifier, expected }
 		}
-		return { ok: true, ...resource, namespace, identifier, expected: undefined }
+		return { ok: false, ...res[0], namespace, identifier: res[1], expected }
 	}
 
 	return undefined
@@ -201,9 +271,7 @@ export const uriBinder: UriBinder = (uris: readonly string[], ctx: UriBinderCont
 
 export function registerCustomResources(config: Config) {
 	for (const [path, res] of Object.entries(config.env.customResources)) {
-		if (res.pack === undefined || res.pack === 'data') {
-			resource(path, { ...res, category: res.category as FileCategory })
-		}
+		resource(path, { ...res, category: res.category as FileCategory })
 	}
 }