+++ packages/java-edition/src/dependency/mcmeta.ts
@@ -1,14 +1,21 @@
 import * as core from '@spyglassmc/core'
-import type { PackMcmeta, ReleaseVersion, VersionInfo } from './common.js'
+import type { TypeDefSymbolData } from '@spyglassmc/mcdoc/lib/binder/index.js'
+import type { PackMcmeta, VersionInfo, VersionInfoReason } from './common.js'
+import { ReleaseVersion } from './common.js'
+
+// DOCS: Update this when a new snapshot cycle begins
+export const NEXT_RELEASE_VERSION = '1.21.4'
 
 /**
  * @param inputVersion {@link core.Config.env.gameVersion}
  */
-export async function resolveConfiguredVersion(
+export function resolveConfiguredVersion(
 	inputVersion: string,
 	versions: McmetaVersions,
-	getPackMcmeta: () => Promise<PackMcmeta | undefined>,
-): Promise<VersionInfo> {
+	packMcmeta: PackMcmeta | undefined,
+	packType: 'assets' | 'data' | undefined,
+	logger: core.Logger,
+): VersionInfo {
 	function findReleaseTarget(version: McmetaVersion): string {
 		if (version.release_target) {
 			return version.release_target
@@ -22,17 +29,20 @@ export async function resolveConfiguredVersion(
 				return versions[i].id
 			}
 		}
-		// DOCS: Update this when a new snapshot cycle begins
-		return '1.21.2'
+		return NEXT_RELEASE_VERSION
 	}
 
-	function toVersionInfo(version: McmetaVersion | undefined): VersionInfo {
+	function toVersionInfo(
+		version: McmetaVersion | undefined,
+		reason: VersionInfoReason,
+	): VersionInfo {
 		version = version ?? versions[0]
 		return {
 			id: version.id,
 			name: version.name,
 			release: findReleaseTarget(version) as ReleaseVersion,
 			isLatest: version === versions[0],
+			reason,
 		}
 	}
 
@@ -40,44 +50,64 @@ export async function resolveConfiguredVersion(
 		throw new Error('mcmeta version list is empty')
 	}
 
+	// This should never happen, but for some reason happens sometimes
+	// https://github.com/SpyglassMC/Spyglass/issues/1621
+	if (inputVersion === undefined) {
+		logger.warn('[resolveConfiguredVersion] Input version was undefined! Falling back to "auto"')
+		inputVersion = 'auto'
+	}
 	inputVersion = inputVersion.toLowerCase()
+
 	versions = versions.sort((a, b) => b.data_version - a.data_version)
 	const latestRelease = versions.find((v) => v.type === 'release')
 	if (inputVersion === 'auto') {
-		const packMcmeta = await getPackMcmeta()
-		if (packMcmeta && latestRelease) {
+		const packFormat = packMcmeta?.pack.pack_format
+		if (packFormat && latestRelease) {
 			// If the pack format is larger than the latest release, use the latest snapshot
-			if (packMcmeta.pack.pack_format > latestRelease.data_pack_version) {
-				return toVersionInfo(versions[0])
+			if (
+				packFormat > (packType === 'assets'
+					? latestRelease.resource_pack_version
+					: latestRelease.data_pack_version)
+			) {
+				return toVersionInfo(versions[0], 'auto')
 			}
 			// Look for versions from recent to oldest, picking the most recent release that matches
 			let oldestRelease = undefined
 			for (const version of versions) {
 				if (version.type === 'release') {
 					// If we already passed the pack format, use the oldest release so far
-					if (packMcmeta.pack.pack_format > version.data_pack_version) {
-						return toVersionInfo(oldestRelease)
+					if (
+						packFormat > (packType === 'assets'
+							? version.resource_pack_version
+							: version.data_pack_version)
+					) {
+						return toVersionInfo(oldestRelease, 'auto')
 					}
-					if (packMcmeta.pack.pack_format === version.data_pack_version) {
-						return toVersionInfo(version)
+					if (
+						packFormat === (packType === 'assets'
+							? version.resource_pack_version
+							: version.data_pack_version)
+					) {
+						return toVersionInfo(version, 'auto')
 					}
 					oldestRelease = version
 				}
 			}
 			// If the pack format is still lower, use the oldest known release version
-			return toVersionInfo(oldestRelease)
+			return toVersionInfo(oldestRelease, 'auto')
 		}
 		// Fall back to the latest release if pack mcmeta is not available
-		return toVersionInfo(latestRelease)
+		return toVersionInfo(latestRelease, 'fallback')
 	} else if (inputVersion === 'latest release') {
-		return toVersionInfo(latestRelease)
+		return toVersionInfo(latestRelease, 'config')
 	} else if (inputVersion === 'latest snapshot') {
-		return toVersionInfo(versions[0])
+		return toVersionInfo(versions[0], 'config')
 	}
 	return toVersionInfo(
 		versions.find((v) =>
 			inputVersion === v.id.toLowerCase() || inputVersion === v.name.toLowerCase()
 		),
+		'config',
 	)
 }
 
@@ -117,7 +147,10 @@ export function getMcmetaSummaryUris(
 	}
 }
 
-export function symbolRegistrar(summary: McmetaSummary): core.SymbolRegistrar {
+export function symbolRegistrar(
+	summary: McmetaSummary,
+	release: ReleaseVersion,
+): core.SymbolRegistrar {
 	const McmetaSummaryUri = 'mcmeta://summary/registries.json'
 
 	/**
@@ -160,6 +193,57 @@ export function symbolRegistrar(summary: McmetaSummary): core.SymbolRegistrar {
 				},
 			)
 		}
+
+		const stateTypes = { block: summary.blocks, fluid: summary.fluids }
+		for (const [type, states] of Object.entries(stateTypes)) {
+			symbols.query(McmetaSummaryUri, 'mcdoc/dispatcher', `mcdoc:${type}_states`)
+				.enter({ usage: { type: 'declaration' } })
+				.onEach(Object.entries(states), ([id, [properties]], query) => {
+					const data: TypeDefSymbolData = {
+						typeDef: {
+							kind: 'struct',
+							fields: Object.entries(properties).map(([propKey, propValues]) => ({
+								kind: 'pair',
+								key: propKey,
+								optional: true,
+								type: {
+									kind: 'union',
+									members: propValues.map(value => ({
+										kind: 'literal',
+										value: { kind: 'string', value },
+									})),
+								},
+							})),
+						},
+					}
+					query.member(id, (stateQuery) => {
+						stateQuery.enter({
+							data: { data },
+							usage: { type: 'declaration' },
+						})
+					})
+				})
+
+			symbols.query(McmetaSummaryUri, 'mcdoc/dispatcher', `mcdoc:${type}_state_keys`)
+				.enter({ usage: { type: 'declaration' } })
+				.onEach(Object.entries(states), ([id, [properties]], query) => {
+					const data: TypeDefSymbolData = {
+						typeDef: {
+							kind: 'union',
+							members: Object.keys(properties).map(propKey => ({
+								kind: 'literal',
+								value: { kind: 'string', value: propKey },
+							})),
+						},
+					}
+					query.member(id, (stateQuery) => {
+						stateQuery.enter({
+							data: { data },
+							usage: { type: 'declaration' },
+						})
+					})
+				})
+		}
 	}
 
 	function addRegistriesSymbols(registries: McmetaRegistries, symbols: core.SymbolUtil) {
@@ -180,8 +264,16 @@ export function symbolRegistrar(summary: McmetaSummary): core.SymbolRegistrar {
 	}
 
 	function addBuiltinSymbols(symbols: core.SymbolUtil) {
-		symbols.query(McmetaSummaryUri, 'loot_table', 'minecraft:empty')
+		if (ReleaseVersion.cmp(release, '1.21.2') < 0) {
+			symbols.query(McmetaSummaryUri, 'loot_table', 'minecraft:empty')
+				.enter({ usage: { type: 'declaration' } })
+		}
+		symbols.query(McmetaSummaryUri, 'model', 'minecraft:builtin/generated')
 			.enter({ usage: { type: 'declaration' } })
+		if (ReleaseVersion.cmp(release, '1.21.4') < 0) {
+			symbols.query(McmetaSummaryUri, 'model', 'minecraft:builtin/entity')
+				.enter({ usage: { type: 'declaration' } })
+		}
 	}
 
 	return (symbols) => {