+++ packages/java-edition/src/mcfunction/parser/argument.ts
@@ -171,7 +171,15 @@ export const argument: mcf.ArgumentParserGetter = (
 		case 'minecraft:entity_summon':
 			return wrap(core.resourceLocation({ category: 'entity_type' }))
 		case 'minecraft:float_range':
-			return wrap(range('float'))
+			return wrap(
+				range(
+					'float',
+					treeNode.properties?.min,
+					treeNode.properties?.max,
+					treeNode.properties?.minSpan,
+					treeNode.properties?.maxSpan,
+				),
+			)
 		case 'minecraft:function':
 			return wrap(core.resourceLocation({ category: 'function', allowTag: true }))
 		case 'minecraft:gamemode':
@@ -181,7 +189,15 @@ export const argument: mcf.ArgumentParserGetter = (
 		case 'minecraft:heightmap':
 			return wrap(core.literal(...HeightmapValues))
 		case 'minecraft:int_range':
-			return wrap(range('integer'))
+			return wrap(
+				range(
+					'integer',
+					treeNode.properties?.min,
+					treeNode.properties?.max,
+					treeNode.properties?.minSpan,
+					treeNode.properties?.maxSpan,
+				),
+			)
 		case 'minecraft:item_enchantment':
 			return wrap(core.resourceLocation({ category: 'enchantment' }))
 		case 'minecraft:item_predicate':
@@ -650,18 +666,24 @@ function range(
 	type: 'float',
 	min?: number,
 	max?: number,
+	minSpan?: number,
+	maxSpan?: number,
 	cycleable?: boolean,
 ): core.Parser<FloatRangeNode>
 function range(
 	type: 'integer',
 	min?: number,
 	max?: number,
+	minSpan?: number,
+	maxSpan?: number,
 	cycleable?: boolean,
 ): core.Parser<IntRangeNode>
 function range(
 	type: 'float' | 'integer',
 	min?: number,
 	max?: number,
+	minSpan?: number,
+	maxSpan?: number,
 	cycleable?: boolean,
 ): core.Parser<FloatRangeNode | IntRangeNode> {
 	const number: core.Parser<core.FloatNode | core.IntegerNode> = type === 'float'
@@ -707,6 +729,21 @@ function range(
 					localize('mcfunction.parser.range.min>max', ans.value[0], ans.value[1]),
 					res,
 				)
+			} else if (minSpan !== undefined || maxSpan !== undefined) {
+				const span = ans.value[0] !== undefined && ans.value[1] !== undefined
+					? Math.abs(ans.value[0] - ans.value[1])
+					: (ans.value[0] ?? ans.value[1] ?? Infinity)
+				if (minSpan !== undefined && span < minSpan) {
+					ctx.err.report(
+						localize('mcfunction.parser.range.span-too-small', span, minSpan),
+						res,
+					)
+				} else if (maxSpan !== undefined && span > maxSpan) {
+					ctx.err.report(
+						localize('mcfunction.parser.range.span-too-large', span, maxSpan),
+						res,
+					)
+				}
 			}
 			return ans
 		},
@@ -1234,7 +1271,14 @@ export function selector(ignoreInvalidPrefix = false): core.Parser<EntitySelecto
 												case 'x_rotation':
 												case 'y_rotation':
 													return core.map<FloatRangeNode>(
-														range('float', undefined, undefined, true),
+														range(
+															'float',
+															undefined,
+															undefined,
+															undefined,
+															undefined,
+															true,
+														),
 														(res, _, ctx) => {
 															if (hasKey(key.value)) {
 																ctx.err.report(
@@ -1353,10 +1397,13 @@ export function scoreHolder(
 	usageType: core.SymbolUsageType,
 	amount: 'multiple' | 'single',
 ): core.Parser<ScoreHolderNode> {
-	return core.map<core.SymbolNode | EntitySelectorNode, ScoreHolderNode>(
-		core.select([{ predicate: (src) => src.peek() === '@', parser: selector() }, {
-			parser: scoreHolderFakeName(usageType),
-		}]),
+	return core.map<core.LiteralNode | core.SymbolNode | EntitySelectorNode, ScoreHolderNode>(
+		core.select([
+			// Technically score holders can start with *, but this doesn't account for it
+			{ prefix: '*', parser: core.literal('*') },
+			{ prefix: '@', parser: selector() },
+			{ parser: scoreHolderFakeName(usageType) },
+		]),
 		(res, _src, ctx) => {
 			const ans: ScoreHolderNode = {
 				type: 'mcfunction:score_holder',
@@ -1366,8 +1413,10 @@ export function scoreHolder(
 
 			if (core.SymbolNode.is(res)) {
 				ans.fakeName = res
-			} else {
+			} else if (EntitySelectorNode.is(res)) {
 				ans.selector = res
+			} else {
+				ans.wildcard = res
 			}
 
 			if (amount === 'single' && ans.selector && !ans.selector.single) {
@@ -1581,27 +1630,80 @@ function vector(options: VectorNode.Options): core.InfallibleParser<VectorNode> 
 	}
 }
 
-const components: core.InfallibleParser<ComponentListNode> = core.map(
-	core.record({
-		start: '[',
-		pair: {
-			key: core.resourceLocation({ category: 'data_component_type' }),
-			sep: '=',
-			value: nbt.parser.entry,
-			end: ',',
-			trailingEnd: true,
-		},
-		end: ']',
-	}),
-	(res) => {
-		const ans: ComponentListNode = {
-			type: 'mcfunction:component_list',
-			range: res.range,
-			children: res.children,
+const components: core.Parser<ComponentListNode> = (src, ctx) => {
+	const release = ctx.project['loadedVersion'] as ReleaseVersion | undefined
+	const allowComponentRemoval = !release || ReleaseVersion.cmp(release, '1.21') >= 0
+
+	const ans: ComponentListNode = {
+		type: 'mcfunction:component_list',
+		range: core.Range.create(src),
+		children: [],
+	}
+
+	if (!src.trySkip('[')) {
+		return core.Failure
+	}
+	src.skipWhitespace()
+
+	while (src.canRead() && src.peek() !== ']') {
+		const start = src.cursor
+		if (allowComponentRemoval && src.tryPeek('!')) {
+			const prefix = core.literal('!')(src, ctx)
+			src.skipWhitespace()
+			const key = core.resourceLocation({ category: 'data_component_type' })(src, ctx)
+			ans.children.push({
+				type: 'mcfunction:component_removal',
+				range: core.Range.create(start, src),
+				children: [prefix, key],
+				prefix,
+				key,
+			})
+			src.skipWhitespace()
+		} else {
+			const key = core.resourceLocation({ category: 'data_component_type' })(src, ctx)
+			src.skipWhitespace()
+			core.literal('=')(src, ctx)
+			src.skipWhitespace()
+			const value = nbt.parser.entry(src, ctx)
+			if (value === core.Failure) {
+				ctx.err.report(
+					localize('expected', localize('parser.record.value')),
+					core.Range.create(src, () => src.skipUntilOrEnd(',', ']', '\r', '\n')),
+				)
+				ans.children.push({
+					type: 'mcfunction:component',
+					range: core.Range.create(start, src),
+					children: [key],
+					key,
+					value: undefined,
+				})
+			} else {
+				ans.children.push({
+					type: 'mcfunction:component',
+					range: core.Range.create(start, src),
+					children: [key, value],
+					key,
+					value,
+				})
+			}
+			src.skipWhitespace()
 		}
-		return ans
-	},
-)
+		if (src.trySkip(',')) {
+			src.skipWhitespace()
+			if (src.peek() === ']') {
+				break
+			}
+		} else {
+			break
+		}
+	}
+
+	src.skipWhitespace()
+	core.literal(']')(src, ctx)
+
+	ans.range.end = src.cursor
+	return ans
+}
 
 const componentTest: core.InfallibleParser<ComponentTestNode> = (src, ctx) => {
 	const start = src.cursor
@@ -1618,6 +1720,7 @@ const componentTest: core.InfallibleParser<ComponentTestNode> = (src, ctx) => {
 	}
 
 	if (src.trySkip('=')) {
+		src.skipWhitespace()
 		const ans: ComponentTestExactNode = {
 			type: 'mcfunction:component_test_exact',
 			range: core.Range.create(start, src),
@@ -1639,6 +1742,7 @@ const componentTest: core.InfallibleParser<ComponentTestNode> = (src, ctx) => {
 	}
 
 	if (src.trySkip('~')) {
+		src.skipWhitespace()
 		if (key.options.category !== undefined) {
 			key.options.category = 'item_sub_predicate_type'
 		}